first school work to make auto class
